+++ src/util/asEdgeTokenId.ts
@@ -1,0 +1,4 @@
+import { asEither, asNull, asString } from 'cleaners'
+
+export type EdgeTokenId = string | null
+export const asEdgeTokenId = asEither(asString, asNull)